refactor(cause-of-death): relocate enum + display meta to shared lib
The frozen-stablecoin obituary block reuses the same CauseOfDeath enum
as DeadStablecoin. Centralizing prevents drift. dead-stablecoins.ts
keeps re-exporting for back-compat.
